[HAPI-] Call the standard endpoints using the new approach
## [HAPI-] Call the standard endpoints using the new approach
