--- src/controllers/api/index.js
@@ -1,5 +1,0 @@
-const healthCheck = require('./health-check');
-
-module.exports = {
-    healthCheck
-};
