--- src/controllers/auth/index.js
@@ -1,7 +1,0 @@
-const register = require('./register');
-const login = require('./login');
-
-module.exports = {
-    register,
-    login    
-};
