+++ src/controllers/auth/login.js
@@ -1,4 +1,4 @@
-const Request = require('@models/RequestAPI');
+const Endpoint = require('@models/settings/Endpoint');
 const User = require('@models/collections/User');
 const Configs = require('@config');
 
@@ -7,36 +7,39 @@ const cookiesConfig = {
     httpOnly: true,
     secure: Configs.secureCookies
 };
-const bodySchema = {
-    email: { type: String, required: true },
-    password: { type: String, required: true }
-};
 
-module.exports = async function (req, res) {
-    try {
-        const request = new Request(req, bodySchema);
-        const body = request.getBody();
-        const user = await User.signIn(body.email, body.password);
-
-        if (user instanceof Error.Log) {
-            let status = 500;
-
-            if (user.name === 'AUTH_INVALID_CREDENTIALS') {
-                status = 401;
+module.exports = new Endpoint({
+    method: 'POST',
+    routePath: '/auth/login',
+    bodySchema: {
+        email: { type: String, required: true },
+        password: { type: String, required: true }
+    },
+    controller: async (req, res) => {
+        try {
+            const body = req.body;
+            const user = await User.signIn(body.email, body.password);
+    
+            if (user instanceof Error.Log) {
+                let status = 500;
+    
+                if (user.name === 'AUTH_INVALID_CREDENTIALS') {
+                    status = 401;
+                }
+    
+                return res.status(status).send(user.toJSON());
             }
-
-            return res.status(status).send(user.toJSON());
+    
+            
+            const response = user.toSession();
+            req.session.currentUser = response;
+    
+            res.cookie('sessionToken', user.token, cookiesConfig);
+            res.cookie('currentUserUID', user._id, cookiesConfig);
+    
+            return res.status(200).send(response.toSuccess().toJSON());
+        } catch(err) {
+            return res.status(500).send(new Error.Log(err).toJSON());
         }
-
-        
-        const response = user.toSession();
-        req.session.currentUser = response;
-
-        res.cookie('sessionToken', user.token, cookiesConfig);
-        res.cookie('currentUserUID', user._id, cookiesConfig);
-
-        return res.status(200).send(response.toSuccess().toJSON());
-    } catch(err) {
-        return res.status(500).send(new Error.Log(err).toJSON());
     }
-}
+});
