+++ src/controllers/collection/delete.js
@@ -1,33 +1,34 @@
 const mongoose = require('mongoose');
 const models = require('../../models');
-const services = require('../../services');
-const CRUD = services.database.crud;
+const CRUD = require('@CRUD');
 const Delete = models.routes.collection.Delete;
-const Request = require('../../models/RequestAPI');
+const Endpoint = require('@src/models/settings/Endpoint');
 const Response = Delete.response;
 
-const bodySchema = {
-    deleteType: { type: String, default: 'one', enum: ['one', 'many'] },
-    collectionName: { type: String, required: true },
-    filter: { type: mongoose.SchemaTypes.Mixed, required: true }, // '_id' || {mongoose-filter}
-    options: { type: Object, default: {} } // mongoose options
-};
-
-module.exports = async function (req, res) {
-    const request = new Request(req, bodySchema);
-    const body = request.getBody();
-
-    try {
-        const deleted = await CRUD.del(body);
-
-        if (deleted){
-            return res.status(200).json(new Response(deleted));
-        } else {
-            const error = new Error.Log('apiResponse.collection.delete');
+module.exports = new Endpoint({
+    method: 'DELETE',
+    routePath: '/collection/delete',
+    bodySchema: {
+        deleteType: { type: String, default: 'one', enum: ['one', 'many'] },
+        collectionName: { type: String, required: true },
+        filter: { type: mongoose.SchemaTypes.Mixed, required: true }, // '_id' || {mongoose-filter}
+        options: { type: Object, default: {} } // mongoose options
+    },
+    controller: async (req, res) => {
+        const body = req.body;
+    
+        try {
+            const deleted = await CRUD.del(body);
+    
+            if (deleted){
+                return res.status(200).json(new Response(deleted));
+            } else {
+                const error = new Error.Log('apiResponse.collection.delete');
+                return res.status(500).json(error.response());
+            }
+        } catch(err) {
+            const error = new Error.Log(err).append('apiResponse.collection.delete');
             return res.status(500).json(error.response());
         }
-    } catch(err) {
-        const error = new Error.Log(err).append('apiResponse.collection.delete');
-        return res.status(500).json(error.response());
     }
-}
+});
