--- src/controllers/collection/get/index.js
@@ -1,7 +1,0 @@
-const queryCollection = require('./queryCollection');
-const doc = require('./doc');
-
-module.exports = {
-    queryCollection,
-    doc
-};
