--- src/controllers/collection/update/index.js
@@ -1,5 +1,0 @@
-const document = require('./document');
-
-module.exports = {
-    document
-};
