--- src/routes/api/index.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-// Controllers
-const healthCheck = require('@controllers/api/health-check');
-
-router.get('/health-check', healthCheck);
-
-module.exports = router;
