--- src/routes/auth/index.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const ctrl = require('@controllers');
-
-router.post('/register', ctrl.auth.register);
-router.post('/login', ctrl.auth.login);
-
-module.exports = router;
