--- src/routes/collection/get.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const ctrl = require('../../controllers');
-
-router.get('/queryCollection', ctrl.collection.get.queryCollection);
-router.get('/doc', ctrl.collection.get.doc);
-
-module.exports = router;
