--- src/routes/index.js
@@ -1,9 +1,0 @@
-const auth = require('./auth');
-const api = require('./api');
-const collection = require('./collection');
-
-module.exports = {
-    auth,
-    api,
-    collection
-};
